desarrollo final

+++ src/main/java/com/example/demo/dtos/TicketDTO.java
@@ -7,13 +7,13 @@ import java.util.List;
 public class TicketDTO {
     private long id;
     private List<ArticleDTO> articles;
-    private double total;
+    private int total;
 
     public TicketDTO(List<ArticleDTO> articles) {
         this.articles = articles;
     }
 
-    public TicketDTO(long id, List<ArticleDTO> articles, double total){
+    public TicketDTO(long id, List<ArticleDTO> articles, int total){
         this.id = id;
         this.total = total;
         this.articles = articles;
